producer/requirements.txt
kafka-python-ng==2.2.2
Faker==19.3.0
python-json-logger==2.0.7
# Ensemble Fraud Detection — Additional Requirements
# Add these to your existing requirements.txt

xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
shap>=0.44.0
scikit-learn>=1.3.0

# GNN Fraud Ring Detection
torch>=2.0.0
torch-geometric>=2.4.0

# Already in your project (keeping for reference)
fastapi
uvicorn
pydantic
mlflow
pandas
numpy
# Agentic AI — Layer 2
langgraph>=0.0.1
langchain>=0.1.0
langchain-openai>=0.0.1
# Redis Feature Store
redis>=5.0.0

# Layer 4: Deep Learning Forecasting
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.0.0

# Layer 5: MLOps & Monitoring
evidently>=0.4.0